requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
importlib-metadata==4.7.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.1
SQLAlchemy==1.4.23
sqlparse==0.4.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
